fatorial

+++ src/main/kotlin/Fatorial.kt
@@ -2,7 +2,7 @@ fun main() {
     print("Digite o numero para fatorar: ")
     val n = readLine()!!.toInt()
 
-    var fatorial = 1
+    var fatorial = 1.toLong()
     var i = n
 
     while (i >= 1){
